requirements.txt
spotipy==2.24.0
langdetect==1.0.9
PyYAML==6.0.2
pytest==8.3.4
python-dotenv==1.0.1
lyricsgenius==3.11.0
